+++ CHANGELOG.md
@@ -1,15 +1,16 @@
 ﻿
 ## Changelog du Gestionnaire de Fond d'Écran
 
-### Version à venir :
+### Version 0.8.2 (3 Mars 2017) :
  - Ajout d'une option pour ajouter l'application au menu contextuel.
  - Amélioration de la fenêtre renommer.
  - Bouton pour acceder au dossier dans l'explorateur de fichiers.
  - Ajout du menu "Note de version" (qui redirige vers les releases de GitHub).
- - Ajout du paramètre : "/D PATH" permettant de démarrer l'application sans passer par la fenêtre de selection du dossier.
+ - Ajout du paramètre "/D PATH" permettant de démarrer l'application sans passer par la fenêtre de selection du dossier.
  - Suppression automatique des points et des espaces dans les parametres des extensions (pour faciliter son utilisation).
  - Recharger un fichier illisible (après sa correction) l'enlève de la liste des mauvais fichiers.
  - Mise en place d'un bouton explorer sur la fenêtre de configuration pour choisir son exécutable externe graphiquement.
+ - Ajout du paramètre /R pour supprimer toutes les données et fichiers du Gestionnaire de Fond d'Écran sur le disque dur et le registre.
 
 ### Version 0.8.1 (28 Février 2017) :
  - Correction de l'affichage lors d'un changement de dossier.